Rename package to deep-unfolding
