+++ .github/workflows/cov.yml
@@ -3,10 +3,10 @@ name: Coverage
 on:
   pull_request:
     branches: [main]
-    paths: [ 'src/unfolding_linear/**', 'tests/**', '.github/workflows/cov.yml' ]
+    paths: [ 'src/deep_unfolding/**', 'tests/**', '.github/workflows/cov.yml' ]
   push:
     branches: [main]
-    paths: [ 'src/unfolding_linear/**', 'tests/**', '.github/workflows/cov.yml' ]
+    paths: [ 'src/deep_unfolding/**', 'tests/**', '.github/workflows/cov.yml' ]
     tags: '*'
   workflow_dispatch:
 
@@ -25,7 +25,7 @@ jobs:
         pip install pip --upgrade
         pip install .[dev]
     - name: Test with pytest and create coverage report
-      run: pytest --cov=unfolding_linear --cov-report=xml
+      run: pytest --cov=deep_unfolding --cov-report=xml
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v3
       with:
