+++ .github/workflows/tests.yml
@@ -3,10 +3,10 @@ name: Tests
 on:
   pull_request:
     branches: [main]
-    paths: [ 'src/unfolding_linear/**', 'tests/**', 'pyproject.toml', '.github/workflows/tests.yml' ]
+    paths: [ 'src/deep_unfolding/**', 'tests/**', 'pyproject.toml', '.github/workflows/tests.yml' ]
   push:
     branches: [main]
-    paths: [ 'src/unfolding_linear/**', 'tests/**', 'pyproject.toml', '.github/workflows/tests.yml' ]
+    paths: [ 'src/deep_unfolding/**', 'tests/**', 'pyproject.toml', '.github/workflows/tests.yml' ]
     tags: '*'
 
 jobs:
